feat: change the way the class work, now it parses all the data in the constructor
change the way the class work, now it parses all the data in the constructor. made the parseing data method private.

BREAKING CHANGE: Now the class parse everything when an object its created. default values for the constructor new files path parameters but the parse method is now private. Before it was parse_from_voc(files...) and now it is _parse_from_voc()